# Optimize BSON codec lookup.