--- a/bson/src/main/org/bson/codecs/BsonArrayCodec.java
+++ b/bson/src/main/org/bson/codecs/BsonArrayCodec.java
@@ -24,6 +24,7 @@ import org.bson.BsonWriter;
 import org.bson.codecs.configuration.CodecRegistry;
 
 import static org.bson.assertions.Assertions.notNull;
+import static org.bson.codecs.BsonValueCodecProvider.getBsonTypeClassMap;
 import static org.bson.codecs.configuration.CodecRegistries.fromProviders;
 
 /**
@@ -34,8 +35,10 @@ import static org.bson.codecs.configuration.CodecRegistries.fromProviders;
 public class BsonArrayCodec implements Codec<BsonArray> {
 
     private static final CodecRegistry DEFAULT_REGISTRY = fromProviders(new BsonValueCodecProvider());
+    private static final BsonTypeCodecMap DEFAULT_BSON_TYPE_CODEC_MAP = new BsonTypeCodecMap(getBsonTypeClassMap(), DEFAULT_REGISTRY);
 
     private final CodecRegistry codecRegistry;
+    private final BsonTypeCodecMap bsonTypeCodecMap;
 
     /**
      * Creates a new instance with a default codec registry that uses the {@link BsonValueCodecProvider}.
@@ -43,7 +46,7 @@ public class BsonArrayCodec implements Codec<BsonArray> {
      * @since 3.4
      */
     public BsonArrayCodec() {
-        this(DEFAULT_REGISTRY);
+        this(DEFAULT_REGISTRY, DEFAULT_BSON_TYPE_CODEC_MAP);
     }
 
     /**
@@ -52,7 +55,12 @@ public class BsonArrayCodec implements Codec<BsonArray> {
      * @param codecRegistry the codec registry
      */
     public BsonArrayCodec(final CodecRegistry codecRegistry) {
-        this.codecRegistry = notNull("codecRegistry", codecRegistry);
+        this(codecRegistry, new BsonTypeCodecMap(getBsonTypeClassMap(), codecRegistry));
+    }
+
+    private BsonArrayCodec(final CodecRegistry codecRegistry, final BsonTypeCodecMap bsonTypeCodecMap) {
+        this.codecRegistry = notNull("Codec registry", codecRegistry);
+        this.bsonTypeCodecMap = notNull("bsonTypeCodecMap", bsonTypeCodecMap);
     }
 
     @Override
@@ -93,7 +101,7 @@ public class BsonArrayCodec implements Codec<BsonArray> {
      * @return the non-null value read from the reader
      */
     protected BsonValue readValue(final BsonReader reader, final DecoderContext decoderContext) {
-        return codecRegistry.get(BsonValueCodecProvider.getClassForBsonType(reader.getCurrentBsonType())).decode(reader, decoderContext);
+        BsonType currentBsonType = reader.getCurrentBsonType();
+        return (BsonValue) bsonTypeCodecMap.get(currentBsonType).decode(reader, decoderContext);
     }
-
 }