This PR introduces performance optimizations for BSON encoding and decoding to  improve efficiency for both BsonArrayCodec BsonDocumentCodec in codec lookups. The latest benchmarks:

**BsonArrayCodec**:
Benchmark | Mode | Count | Before (ops/s) | After (ops/s) |
-- | -- | -- | -- | --
BsonArrayCodecBenchmark.decode | thrpt | 60 | 29,222.3 | 43,156.9 
BsonArrayCodecBenchmark.encode | thrpt | 60 | 12,611.6 | 14,886.9 

**BsonDocumentCodec**:
 The BsonDocumentBenchmark  parses simple BsonInt32 fields to minimize the impact of other codecs to evaluate BsonDocumentCodec's internal performance.

Benchmark | Mode | Count | Before (ops/s) | After (ops/s) | 
-- | -- | -- | -- | --
BsonDocumentBenchmark.decode | thrpt | 60 | 30,511.05 | 41,312.66 
BsonDocumentBenchmark.encode | thrpt | 60 | 33,476.20 | 39,243.81



Changes:
- Use `BsonTypeCodecMap` for encoding and decoding in `BsonArrayCodec` and `BsonDocumentCodec` to enable faster access.
- Optimize `BsonTypeClassMap` by replacing `Map` with a plain array to eliminate redundant hash computations.

JAVA-5339